refactor(pa): Use FlyDSL autotune persistence
