--- kernels/attention/pa_metadata_grid_tuning.csv
@@ -1,19 +1,0 @@
-arch,num_cu,batch_size,num_blocks,context_length,query_length,per_token_kv,num_query_heads,num_kv_heads,head_dim,block_size,grid_multiplier,latency_us
-gfx942,80,1,1,1024,1,0,16,1,128,1024,1,11.440
-gfx942,80,3,6,1027,4,1,16,1,128,1024,1,21.160
-gfx942,80,8,256,32768,4,1,16,1,128,1024,1,122.200
-gfx942,80,16,64,4096,1,0,16,1,128,1024,1,20.680
-gfx942,80,16,128,8192,4,0,16,1,128,1024,2,58.801
-gfx942,80,16,128,8192,4,1,16,1,128,1024,1,75.000
-gfx942,80,32,256,8192,1,0,16,1,128,1024,3,35.520
-gfx942,80,64,256,4096,1,1,16,1,128,1024,3,38.360
-gfx942,80,64,256,4096,2,0,16,1,128,1024,2,53.600
-gfx942,80,64,256,4096,2,1,16,1,128,1024,2,59.920
-gfx942,80,64,256,4096,3,0,16,1,128,1024,2,71.401
-gfx942,80,64,256,4096,3,1,16,1,128,1024,2,81.800
-gfx942,80,64,512,8192,1,0,16,1,128,1024,3,49.520
-gfx942,80,81,648,8192,1,0,16,1,128,1024,3,65.840
-gfx942,80,81,648,8192,4,0,16,1,128,1024,2,184.720
-gfx942,80,81,648,8192,4,1,16,1,128,1024,1,263.400
-gfx942,80,128,1024,8192,4,1,16,1,128,1024,1,415.640
-gfx942,80,256,512,2048,4,1,16,1,128,1024,1,232.000
